Add OnigScanner::enableTracing()
This is part of the effort to address https://github.com/atom/atom/issues/979.

enableTracing() asks the scanner to track how long each regexp
matching took and report slow ones back as a part of findNextToken() result.
With this data, syntax highlighting module author will be able to identify
the slow regexps and make it faster.